Merge noteable from Tyler (#67)
* Adding a new page "noteables" 

I want a page for external links to help with SEO. (Pages I want to rise higher in rankings that are not tylervawser.com)
I was going to title it "links" but figure that would create problems with the code and arrays. 
Instead I named it noteables. I tried to keep everything consistent with favorites and renamed the arrays, etc.

* Rename noteables.js to noteable.js

* Updated twitter link to include intent follow

* Update noteable.js

Co-authored-by: Tyler Vawser <TVawser@gmail.com>